requirements.txt
django==1.9.8
django-crispy-forms==1.6.1
django-formtools==1.0
django-pure-pagination==0.3.0
django-simple-captcha==0.4.6
httplib2==0.9.2
MySQL-python==1.2.5
olefile==0.43
Pillow==4.0.0
six==1.10.0